Remove require try
vmdb-logger should not require any active-support files so
that it can be included in code that does not (and cannot)
use active-support.

Currently the VixDiskLibServer runs in a very vanilla environment
so as to not run into any conflicts with versions of SSL required
by the VMWare VixDiskLib (vddk). This change allows vmdb-logger
to provide this support.


(transferred from manageiq-gems-pending@f67d7a01d11d28c7520c735bc62a1c8a712eaa5d)