Implements @local_level to vmdb-logger
Rails makes use of a module included in their default logger called
`ActiveSupport::LoggerThreadSafeLevel`.  This basically adds a instance
variable to the logger called `@local_level` that sits in front of the
`level` method.  When set explicitly, it will make sure that the logging
level set specifically for that thread is honored.

Since we are currently using a threaded websever, puma, this allows us
to make use of setting local levels for individual threads that don't
affect the other processes and mess with the global logging level state.

In activesupport, this is implemented using the gem, concurent-ruby,
implementation of Concurrent::Map, which basically adds a mutex around a
regular ruby Hash when adding or removing data (on MRI specifically).
Since we currently aren't using any non-MRI version of ruby to deploy
ManageIQ, this was implemented using vanilla ruby instead.  If we wish
to support JRuby and other ruby interpreters, using concurrent ruby
should be considered (we already have the the gem when using rails).


(transferred from manageiq-gems-pending@e2bdbfbff746ca9b2b4f559b11c0d498b5f811af)