Remove require outside of test, let initialize do it
Today I learned that filtering specs by "linux" still runs code outside of
tests. Code within the describe block is still evaluated before the tests
within it are excluded.  The require needs to be done within a let, before, or
an example. Instead, we can let the initialize do the require and make sure
that is run first by changing it to let!.